Added conditonal rendering and kept the desktop and mobile its very own layouts. Test data to make sure all inputs work.
