+++ client/src/pages/DesktopLayout.css
@@ -1,0 +1,74 @@
+.desktop-container {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  height: 100%;
+  padding-left: 101px;
+}
+
+.desktop-row {
+  width: 100%;
+  display: flex;
+}
+
+.desktop-col {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+.desktop-col-half {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100vh;
+}
+
+.left-content {
+  padding: 40px 20px;
+}
+
+.right-content {
+  width: 40%;
+  background-color: #112423;
+  border-radius: 22px 0 0 22px;
+  padding: 40px 20px;
+}
+
+.journal-card {
+  width: 360px;
+  height: 390px;
+  background-color: #0a1514;
+  color: white;
+  padding: 32px 20px;
+  border-radius: 24px;
+  border: none;
+}
+
+.plant {
+  width: 200px;
+  height: 200px;
+  margin-bottom: 15px;
+  outline: none;
+}
+
+.bar {
+  max-width: 500px;
+  height: 12px;
+  background-color: white;
+  margin: 5px 0 0;
+}
+
+.bg-fill-bar {
+  background-color: #1b998d;
+}
+
+.line {
+  width: 100%;
+  height: 1px;
+  background-color: grey;
+}
+
+.padding {
+  margin: 0 10px;
+}
