+++ client/src/pages/MobileLayout.css
@@ -105,7 +105,6 @@ button {
   padding: 10px 0;
 }
 
-/* Temporary */
 .emojis {
   width: 100%;
   height: 34px;
